APP-14871: fix lint gates and split client reporting into its own file
- rewrite furthestPendingReport without a computed slice index (silences a
  gosec G602 false positive) and //nolint:nilnil the test fake's NewStream.
- move the client-side connection-reporting subsystem (collector, send/flush,
  classify*, sdkVersion, dialDurationMS) out of wrtc_client.go into
  wrtc_client_report.go, pairing it with wrtc_client_report_test.go. Pure code
  move, no behavior change.
